+++ src/test/java/StatsServiceTest.java
@@ -10,9 +10,9 @@ public class StatsServiceTest {
     public void sumAllSales() {                                       // общая сумма всех продаж
         StatsService service = new StatsService();
 
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18}; // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18}; // продажи по месяцам
         int expectedSales = 180;                                      // общая сумма всех продаж
-        int allSum = service.sumAllSales(sales);
+        long allSum = service.sumAllSales(sales);
 
         Assertions.assertEquals(expectedSales, allSum);
     }
@@ -22,9 +22,9 @@ public class StatsServiceTest {
     public void sumAverageSalesInMonth() {                               // средняя сумма продаж в месяц
         StatsService service = new StatsService();
 
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};    // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};    // продажи по месяцам
         int expectedSumAverageSalesInMonth = 15;                        // средняя сумма продаж в месяц
-        int Average = service.sumAverageSalesInMonth(sales);
+        long Average = service.sumAverageSalesInMonth(sales);
 
         Assertions.assertEquals(expectedSumAverageSalesInMonth, Average);
 
@@ -35,9 +35,9 @@ public class StatsServiceTest {
     public void sumMaxSalesMonth() {                                   // Номер месяца, в котором был пик продаж
         StatsService service = new StatsService();
 
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18}; // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18}; // продажи по месяцам
         int expectedSumMaxSalesMonth = 8;                             // Номер месяца, в котором был пик продаж
-        int SumMaxSalesMonth = service.sumMaxSalesMonth(sales);
+        long SumMaxSalesMonth = service.sumMaxSalesMonth(sales);
 
         Assertions.assertEquals(expectedSumMaxSalesMonth, SumMaxSalesMonth);
 
@@ -46,9 +46,9 @@ public class StatsServiceTest {
     @Test
     public void sumMinSalesMonth() {                                    // Номер месяца, в котором был минимум продаж
         StatsService service = new StatsService();
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};   // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};   // продажи по месяцам
         int expectedSumMinSalesMonth = 9;                              // Номер месяца, в котором был минимум продаж
-        int sumMinSalesMonth = service.sumMinSalesMonth(sales);
+        long sumMinSalesMonth = service.sumMinSalesMonth(sales);
 
         Assertions.assertEquals(expectedSumMinSalesMonth, sumMinSalesMonth);
 
@@ -58,9 +58,9 @@ public class StatsServiceTest {
     @Test
     public void minCalcMonthBelowAverage() {                             // Количество месяцев, в которых продажи были ниже среднего
         StatsService service = new StatsService();
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};    // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};    // продажи по месяцам
         int expectedMinCalcMonthBelowAverage = 5;                       // Количество месяцев, в которых продажи были ниже среднего
-        int minCalcMonthBelowAverage = service.minCalcMonthBelowAverage(sales);
+        long minCalcMonthBelowAverage = service.minCalcMonthBelowAverage(sales);
 
         Assertions.assertEquals(expectedMinCalcMonthBelowAverage, minCalcMonthBelowAverage);
 
@@ -70,9 +70,9 @@ public class StatsServiceTest {
     @Test
     public void maxCalcMonthHigherAverage() {                               // Количество месяцев, в которых продажи были выше среднего
         StatsService service = new StatsService();
-        int sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};       // продажи по месяцам
+        long sales[] = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};       // продажи по месяцам
         int expectedMaxCalcMonthHigherAverage = 5;                          // Количество месяцев, в которых продажи были выше среднего
-        int maxCalcMonthHigherAverage = service.maxCalcMonthHigherAverage(sales);
+        long maxCalcMonthHigherAverage = service.maxCalcMonthHigherAverage(sales);
 
         Assertions.assertEquals(expectedMaxCalcMonthHigherAverage, maxCalcMonthHigherAverage);
